refactor/utils
