+++ src/print.js
@@ -4,13 +4,18 @@
 const chalk = require('chalk');
 const figlet = require('figlet');
 
+const settings = {
+  horizontalLayout: 'full',
+};
+
+const START_MESSAGE = 'Bulls And Cows';
+const WIN_MESSAGE = 'Congratulations';
+
 module.exports = {
   start: function() {
     console.log(
       chalk.bgMagenta(
-        figlet.textSync('Bulls And Cows', {
-          horizontalLayout: 'full',
-        })
+        figlet.textSync(START_MESSAGE, settings)
       )
     );
   },
@@ -20,9 +25,7 @@ module.exports = {
   win: function() {
     console.log(
       chalk.bgMagenta(
-        figlet.textSync('Congratulations', {
-          horizontalLayout: 'full',
-        })
+        figlet.textSync(WIN_MESSAGE, settings)
       )
     );
   },
